requirements.txt
mysql-connector-python
flask
flask-wtf